+++ .eslintrc.js
@@ -22,6 +22,7 @@ module.exports = {
   rules: {
     '@typescript-eslint/ban-types': 1, // StyledComponentBase<any, {}> failed every time
     '@typescript-eslint/no-empty-function': 0,
+    '@typescript-eslint/no-inferrable-types': 1,
     'arrow-body-style': 0,
     'arrow-parens': 0,
     'comma-dangle': 1,
@@ -30,7 +31,7 @@ module.exports = {
     'function-paren-newline': 0,
     'implicit-arrow-linebreak': 0,
     'import/extensions': 0,
-    'import/no-extraneous-dependencies': 1,
+    'import/no-extraneous-dependencies': 0,
     'import/no-unresolved': 0,
     'import/prefer-default-export': 1,
     'jsx-a11y/label-has-associated-control': 0,