+++ .storybook/decorators.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { FoundryProvider } from '../src/context';
+
+export const withFoundryContext = Story => (
+  <FoundryProvider>
+    <Story />
+  </FoundryProvider>
+);