+++ src/components/Modal/Modal.stories.tsx
@@ -1,11 +1,12 @@
-import React, { useState } from 'react';
+import React, { useCallback, useRef, useState } from 'react';
 import styled from 'styled-components';
 import { Story, Meta } from '@storybook/react';
 
 import colors from '../../enums/colors';
-import Modal, { ModalProps } from './Modal';
+import Modal, { ModalProps, ModalApi } from './Modal';
 import Button from '../Button/Button';
 import Card from '../Card';
+import { withFoundryContext } from '../../../.storybook/decorators';
 
 const Background = styled.div`
   background-image: url(https://source.unsplash.com/weekly?landscape);
@@ -37,6 +38,14 @@ export const Default: Story<DefaultProps> = ({
     onClose();
   };
 
+  const modalRef = useRef<ModalApi>();
+
+  const customCloseButtonClick = useCallback(() => {
+    if (modalRef.current) {
+      modalRef.current.close();
+    }
+  }, [modalRef]);
+
   return (
     <Background>
       <Card elevation={1} header="Use this button to open the modal again">
@@ -54,11 +63,12 @@ export const Default: Story<DefaultProps> = ({
           backgroundBlur={`${backgroundBlur}rem`}
           onClickOutside={onClickOutside ? handleClose : undefined}
           onClose={handleClose}
+          ref={modalRef}
         >
           <Card
             header="Hello world!"
             footer={
-              <Button color={colors.primaryDark} onClick={handleClose}>
+              <Button color={colors.primaryDark} onClick={customCloseButtonClick}>
                 Okay...
               </Button>
             }
@@ -97,6 +107,7 @@ export default {
       control: { type: 'range', min: 0, max: 5, step: 0.1 },
     },
   },
+  decorators: [withFoundryContext],
   parameters: {
     layout: 'fullscreen',
     design: {