+++ src/components/RangeSlider/RangeSlider.tsx
@@ -20,8 +20,8 @@ import {
   SelectedRangeProps,
   DomainLabelProps,
 } from './types';
-import { useAnalytics, useTheme } from '../../context';
-import { Div } from '../../htmlElements';
+import { useAccessibilityPreferences, useAnalytics, useTheme } from '../../context';
+import { StyledBaseDiv } from '../../htmlElements';
 
 export const Container = styled.div`
   ${({ showDomainLabels, hasHandleLabels, disabled, beingDragged = false }: ContainerProps) => `
@@ -133,7 +133,7 @@ export const SlideRail = styled.div`
 `;
 
 export const SelectedRangeRail = styled.div`
-  ${({ min, max, selectedRange }: SelectedRangeProps) => {
+  ${({ min, max, selectedRange, animateRangeRail }: SelectedRangeProps) => {
     const { colors } = useTheme(); // TODO: don't force the color to be primary
     return `
       position: absolute;
@@ -142,7 +142,13 @@ export const SelectedRangeRail = styled.div`
       left: ${((selectedRange[0] - min) / (max - min)) * 100}%;
       right: ${((max - selectedRange[1]) / (max - min)) * 100}%;
 
+      ${
+        animateRangeRail
+          ? `
       transition: left .3s, right .3s;
+      `
+          : ''
+      }
 
       background-color: ${colors.primary};
     `;
@@ -162,7 +168,7 @@ export const DomainLabel = styled.div`
   }}
 `;
 
-export const Marker = styled(Div)`
+export const Marker = styled(StyledBaseDiv)`
   ${({ sliderPosition = 0 }) => {
     const { colors } = useTheme();
     return `
@@ -177,7 +183,7 @@ export const Marker = styled(Div)`
     `;
   }}
 `;
-export const MarkerLabel = styled(Div)`
+export const MarkerLabel = styled(StyledBaseDiv)`
   ${({ color }) => {
     const { colors } = useTheme();
     return `
@@ -238,6 +244,7 @@ export const RangeSlider = ({
   testId,
 }: RangeSliderProps): JSX.Element | null => {
   const { colors } = useTheme();
+  const { prefersReducedMotion } = useAccessibilityPreferences();
   const hasHandleLabels = useRef(false);
   const initializing = useRef(true);
 
@@ -272,12 +279,16 @@ export const RangeSlider = ({
 
   const selectedRange = [
     Math.min(
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       ...processedValues.map((val: number | ValueProp) =>
         typeof val === 'number' ? val : val.value,
       ),
       showSelectedRange && values && values.length === 1 ? min : Infinity,
     ),
     Math.max(
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       ...processedValues.map((val: number | ValueProp) =>
         typeof val === 'number' ? val : val.value,
       ),
@@ -302,9 +313,11 @@ export const RangeSlider = ({
     [handleEventWithAnalytics, onDrag, containerProps],
   );
 
+  const finalDebounceInterval = prefersReducedMotion ? 0 : debounceInterval;
+
   // set the drag value asynchronously at a lower frequency for better performance
   const valueBuffer = useRef(0);
-  const debouncedDrag = debounce(() => handleDrag(valueBuffer.current), debounceInterval);
+  const debouncedDrag = debounce(() => handleDrag(valueBuffer.current), finalDebounceInterval);
   const blurRef = useRef(null);
 
   // keep track of which handle is being dragged (if any)
@@ -319,10 +332,12 @@ export const RangeSlider = ({
   });
 
   // get the x offset and an animation setter function
-  const [{ x, y }, set] = useSpring(
-    () => ({ to: { x: pixelPositions[0], y: 0 }, config: { friction: 13, tension: 100 } }),
-    [values],
-  );
+  const [{ x, y }, set] = useSpring(() => ({
+    to: { x: pixelPositions[0], y: 0 },
+    friction: 13,
+    tension: 100,
+    immediate: prefersReducedMotion,
+  }));
 
   const handleSlideRailClick = useCallback(
     (e: React.MouseEvent) => {
@@ -386,7 +401,7 @@ export const RangeSlider = ({
         x: down ? deltaX : pixelPositions[0],
         y: down ? deltaY : 0,
 
-        immediate: springOnRelease ? down : true,
+        immediate: prefersReducedMotion || springOnRelease ? down : true,
         config: { friction: 13, tension: 100 },
       });
     },
@@ -411,14 +426,14 @@ export const RangeSlider = ({
 
       // always snap to position on initial render
       // then leave snapping up to springOnRelease
-      immediate: !springOnRelease || initializing.current,
+      immediate: prefersReducedMotion || !springOnRelease || initializing.current,
       config: { friction: 13, tension: 100 },
       onRest: () => {
         // wait for the first "set" to finish before turning off immediate mode
         initializing.current = false;
       },
     });
-  }, [pixelPositions, set, springOnRelease]);
+  }, [pixelPositions, set, springOnRelease, prefersReducedMotion]);
 
   return (
     <StyledContainer
@@ -441,6 +456,7 @@ export const RangeSlider = ({
             max={max}
             values={processedValues}
             selectedRange={selectedRange}
+            animatedRangeRail={!prefersReducedMotion}
             ref={selectedRangeRailRef}
             {...selectedRangeRailProps}
           />