+++ src/index.ts
@@ -12,9 +12,26 @@ import Table from './components/Table';
 import Tag from './components/Tag';
 import Text from './components/Text';
 import TextInput from './components/TextInput';
+import Toggle from './components/Toggle';
 import Progress from './components/Progress';
 import Skeleton from './components/Skeleton';
-import { FoundryProvider, FoundryContext, useTheme } from './context';
+import { FoundryProvider, FoundryContext, useTheme, withGlobalStyle } from './context';
+import { useStateWithPrevious, useWindowSizeObserver, useScrollObserver } from './utils/hooks';
+import { clamp } from './utils/math';
+
+import {
+  StyledBaseDiv,
+  StyledBaseSpan,
+  StyledBaseButton,
+  StyledBaseInput,
+  StyledBaseLabel,
+  StyledBaseHR,
+  StyledBaseTable,
+  StyledBaseTH,
+  StyledBaseTD,
+  StyledBaseTR,
+  StyledBaseTextArea,
+} from './htmlElements';
 
 import colors from './enums/colors';
 import timings from './enums/timings';
@@ -22,8 +39,28 @@ import fonts from './enums/fonts';
 import variants from './enums/variants';
 import feedbackTypes from './enums/feedbackTypes';
 import checkboxTypes from './enums/checkboxTypes';
+import {
+  disabledStyles,
+  getBackgroundColorFromVariant,
+  getFontColorFromVariant,
+} from './utils/color';
+import { useReducedMotion } from './utils/a11y';
 
 export {
+  // Globally styled base elements
+  StyledBaseDiv,
+  StyledBaseSpan,
+  StyledBaseButton,
+  StyledBaseInput,
+  StyledBaseLabel,
+  StyledBaseHR,
+  StyledBaseTable,
+  StyledBaseTH,
+  StyledBaseTD,
+  StyledBaseTR,
+  StyledBaseTextArea,
+  withGlobalStyle,
+  // Components
   Button,
   Card,
   Checkbox,
@@ -39,17 +76,28 @@ export {
   Tag,
   Text,
   TextInput,
-  /**
-   * @deprecated The Progress loading skeleton is being replaced by the Skeleton component
-   */
-  Progress,
+  Toggle,
   FoundryProvider,
   FoundryContext,
+  // Utils and helpers
   useTheme,
+  clamp,
+  getFontColorFromVariant,
+  getBackgroundColorFromVariant,
+  disabledStyles,
+  useStateWithPrevious,
+  useWindowSizeObserver,
+  useScrollObserver,
+  useReducedMotion,
+  // Defaults
   colors,
   timings,
   fonts,
+  // Enums and types
   variants,
   feedbackTypes,
   checkboxTypes,
+  // deprecated exports
+  /** @deprecated The Progress loading skeleton is being replaced by the Skeleton component */
+  Progress,
 };